[BAD-601]: The bunch of unit tests for pentaho-hadoop-shims (master branch) are failed on windows
What was done:
1)Tests that use local file system for manipulations with files are located in the separate class.
2)These tests for cdh, hdp, emr and hdi shims are run successfully on Windows;
3)These tests are skipped for mapr510 and mapr520 shims on Windows.
